プルリクの修正

+++ apps/kita/database/migrations/2023_06_30_083026_create_members_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up()
     {
         Schema::create('members', function (Blueprint $table) {
-            $table->unsignedBigInteger('id')->autoIncrement();
+            $table->unsignedInteger('id')->autoIncrement();
             $table->string('name');
             $table->string('email')->unique();
             $table->string('password');
